Safari-609.1.11.1/LayoutTests/accessibility/noscript-ignored-expected.txt

This tests that elements inside noscript are ignored when making the accessible canvas children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canvas.childrenCount is 0
PASS successfullyParsed is true

TEST COMPLETE
